Apply  feedback from  code review
