+++ src/main/kotlin/com/example/idolkingdom/model/BloodType.kt
@@ -1,0 +1,5 @@
+package com.example.idolkingdom.model
+
+enum class BloodType {
+    A, B, AB, O
+}
