+++ src/main/kotlin/com/example/idolkingdom/model/CommentTarget.kt
@@ -1,0 +1,6 @@
+package com.example.idolkingdom.model
+
+data class CommentTarget(
+    val idolGroup: IdolGroup,
+    val school: School
+)
