+++ src/main/kotlin/com/example/idolkingdom/model/Entertainment.kt
