+++ src/main/kotlin/com/example/idolkingdom/model/Location.kt
