+++ src/main/kotlin/com/example/idolkingdom/model/SchoolLevel.kt
@@ -1,0 +1,5 @@
+package com.example.idolkingdom.model
+
+enum class SchoolLevel {
+    ELEMENT, MIDDLE, HIGH, UNIVERSAL
+}
