+++ src/main/kotlin/com/example/idolkingdom/model/User.kt
@@ -1,21 +1,14 @@
 package com.example.idolkingdom.model
 
-import java.util.*
+import java.time.ZonedDateTime
 
 data class User(
     val id: Int,
+    val email: String,
+    val password: String,
     val name: String,
     val school: School,
-    val account: Account,
     val lovedGroups: List<IdolGroup>,
     val votedList: List<Vote>,
-    val createdAt: Date
-) {
-
-    data class Account(
-        val email: String,
-        val password: String
-//        val kakaoId: String,
-//        val facebookId: String
-    )
-}
+    val createdAt: ZonedDateTime
+)
