+++ src/main/kotlin/com/example/idolkingdom/model/Voting.kt
@@ -1,0 +1,9 @@
+package com.example.idolkingdom.model
+
+import java.util.*
+
+data class Voting(
+    val id: Int,
+    val user: User,
+    val date: Date
+)
